--- a/Source/MQTTnet.AspnetCore/Client/Tcp/BufferExtensions.cs
+++ b/Source/MQTTnet.AspnetCore/Client/Tcp/BufferExtensions.cs
@@ -1,7 +1,10 @@
-﻿using System;
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
 using System.Runtime.InteropServices;
 
-namespace MQTTnet.AspNetCore.Client.Tcp
+namespace Microsoft.AspNetCore.Server.Kestrel.Transport.Sockets.Internal
 {
     public static class BufferExtensions
     {