--- a/Source/MQTTnet.AspnetCore/Client/Tcp/SocketSenderReceiverBase.cs
+++ b/Source/MQTTnet.AspnetCore/Client/Tcp/SocketSenderReceiverBase.cs
@@ -1,0 +1,23 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.IO.Pipelines;
+using System.Net.Sockets;
+
+namespace Microsoft.AspNetCore.Server.Kestrel.Transport.Sockets.Internal
+{
+    public abstract class SocketSenderReceiverBase : IDisposable
+    {
+        protected readonly Socket _socket;
+        protected readonly SocketAwaitableEventArgs _awaitableEventArgs;
+
+        protected SocketSenderReceiverBase(Socket socket, PipeScheduler scheduler)
+        {
+            _socket = socket;
+            _awaitableEventArgs = new SocketAwaitableEventArgs(scheduler);
+        }
+
+        public void Dispose() => _awaitableEventArgs.Dispose();
+    }
+}