hi there I cant stop fiddling with perf so here is another attempt to improve:

counters from MessageProcessingMqttConnectionContextBenchmark

before:

|             Method |     Mean |     Error |    StdDev |     Gen 0 |    Gen 1 | Allocated |
|-------------------- |---------:|----------:|----------:|----------:|---------:|----------:|
| Send_10000_Messages | 16.50 ms | 0.3514 ms | 0.7098 ms | 1890.6250 | 140.6250 |   7.94 MB |

after:

|              Method |     Mean |     Error |    StdDev |     Gen 0 |    Gen 1 | Allocated |
|-------------------- |---------:|----------:|----------:|----------:|---------:|----------:|
| Send_10000_Messages | 9.188 ms | 0.1794 ms | 0.2515 ms | 1281.2500 | 156.2500 |   5.52 MB |

main changes are:
- added body reader abstraction so the asp.net part can use its own implementation 
- use spans for reading body to avoid byte[] copy SpanBasedMqttPacketBodyReader 
- reuse body reader
- added new receive overload and provide packets with the packetcompleted event. that way the reader can read many packets on block if they are present in the buffer

x dont flush every packet to the underlying buffers so there are less network calls in total
this didnt work

if the throughput scales as well in reality im at 200k msgs/sec :)

let me know what you think