@JanEggers 2018-08-08
added netcoreapp compilation,that gives a nice additional boost:

|              Method |     Mean |     Error |    StdDev |     Gen 0 |   Gen 1 | Allocated |
|-------------------- |---------:|----------:|----------:|----------:|--------:|----------:|
| Send_10000_Messages | 6.868 ms | 0.1353 ms | 0.2606 ms | 1617.1875 | 31.2500 |   1.76 MB |