@chkr1011 2018-09-13
Hi, 
I have no time for a review at the moment. But don't worry I will review it in a few days 😄 
Could you please have a look at the MQTTv5 branch? I changed the way how the serializer is choosen. Maybe your change requires some changes to support it. The idea is to read the first part of the _Connect_ packet to determine the serializer which should be choosen. Then the data is reverted and the serializer deserialized this initial part again to have a valid packet (for unit tests etc.)
Best regards
Christian