@JanEggers 2018-09-13
Hi Chirtian, no worries... this isnt important right now Im just tinkering. I will try to rebase my changes as soon as im having time for it.