@JanEggers 2018-09-13
I had a look at your changes. let me know when u integrt that in dev branch. I will rebase and adjust my code afterwards