@JanEggers 2018-12-22
@chkr1011: can you please also merge the mqtt v5 stuff you did into dev so i can do another rebase