@JanEggers 2018-12-22
and your code signing cert is no longer valid thats why the builds fail...