@JanEggers 2018-12-22
@chkr1011 thx again christian, please do not yet review or merge, the code is working with the mqtt v5 changes but there is a negative perf impact somewhere.

I get 20 ms / 10000 packets on net461 and 11 ms on netcore before I did the changes It took 13 on net461 and 6 on netcore