@chkr1011 2018-12-22
OK let me know as soon as I can have a review 😄  This new Span etc. stuff looks very interesting. Do you have a recommendation for a tutorial etc.? Everything I found is slightly theoretical and thus I don't understand it 😄