@JanEggers 2018-12-22
have a look at:

https://adamsitnik.com/Span/

for me the most important method is slice as it provides access to a sub range without allocating a copy of that range.